// src/types/@gearbox-protocol/core-v2/contracts/interfaces/IPhantomERC20.ts
export interface TransferEventObject {
  from: string;
  to: string;
  value: BigNumber;
}
export type TransferEvent = TypedEvent<
  [string, string, BigNumber],
  TransferEventObject
>;